+++ index.js
@@ -17,18 +17,27 @@ program
   .requiredOption('-p, --path <path>', 'path/url to swagger scheme')
   .option('-o, --output <output>', 'output path of typescript api file', './')
   .option('-n, --name <name>', 'name of output typescript api file', 'Api.ts')
+  .option('-d, --default-as-success', 'use "default" response status code as success response too.\n' + 'some swagger schemas have "default" response status code.', false)
   .option('--route-types', 'generate type definitions for API routes', false)
   .option('--no-client', 'do not generate an API class', false);
  
 program.parse(process.argv);
 
-const { path, output, name, routeTypes, client } = program;
+const {
+  path,
+  output,
+  name,
+  routeTypes,
+  client,
+  defaultAsSuccess,
+} = program;
 
 generateApi({
   name,
   url: path,
   generateRouteTypes: routeTypes,
   generateClient: client,
+  defaultResponseAsSuccess: defaultAsSuccess,
   input: resolve(process.cwd(), path),
   output: resolve(process.cwd(), output || '.')
 })